changes in updateExhibition for profilePic

+++ src/api/controllers/exhibition.controller.js
@@ -2,6 +2,7 @@ import { exhibitionServices } from "../services/exhibition/exhibition.service.js
 import fs from "fs";
 
 export const setExhibition = async (req, res, next) => {
+  if(!req.file) return res.status(404).send("file not selected")
   try {
     const data = await exhibitionServices.createProfile({
       ...req.body,
@@ -17,15 +18,19 @@ export const setExhibition = async (req, res, next) => {
 export const updateExhibition = async (req, res, next) => {
   const id = req.params;
   try {
-    const exbData = await exhibitionServices.fetchProfile(id);
+    if (req.file) {
+      const exbData = await exhibitionServices.fetchProfile(id);
 
-    if (fs.existsSync(exbData.imageURL)) fs.unlinkSync(exbData.imageURL);
-    const updatedExb = await exhibitionServices.updateProfile(id, {
-      ...req.body,
-      imageURL: req.file.path,
-    });
-
-    return res.status(201).send({ success: true, updatedExb });
+      if (fs.existsSync(exbData.imageURL)) fs.unlinkSync(exbData.imageURL);
+      const updatedExb = await exhibitionServices.updateProfile(id, {
+        ...req.body,
+        imageURL: req.file.path,
+      });
+      return res.status(201).send({ success: true, updatedExb });
+    } else {
+      const updatedExb = await exhibitionServices.updateProfile(id, req.body);
+      return res.status(201).send({ success: true, updatedExb });
+    }
   } catch (error) {
     console.log(error);
     res.status(501).send({ success: false, message: error.message });
